Move MultiDomainBasicAuth to networking.auth
